split rwp manifest code into helpers package
